moved activity view
