+++ app/src/main/java/com/globant/weatherapp/mvp/view/WeatherView.kt
@@ -2,6 +2,7 @@ package com.globant.weatherapp.mvp.view
 
 import android.app.Activity
 import com.globant.weatherapp.mvp.contracts.WeatherContracts
+import com.globant.weatherapp.mvp.view.base.ActivityView
 
 class WeatherView(activity: Activity): ActivityView(activity), WeatherContracts.View {
